+++ Gemfile
@@ -1,0 +1,4 @@
+source 'https://rubygems.org'
+
+# bundle robocop.
+gem 'rubocop', '>= 1.0', '< 2.0'
