--- app/Student.rb
@@ -1,12 +1,0 @@
-require_relative 'Person'
-
-class Student < Person
-  def initialize(name*, age, parent_permission*, classroom)
-    super(name, age, parent_permission)
-    @classroom = classroom
-  end
-
-  def play_hooky
-    puts "¯\(ツ)/¯";
-  end
-end
